requirements.txt
asn1==2.5.0
cffi==1.15.0
cryptography==36.0.2
enum-compat==0.0.3
pycparser==2.21
PyJWT==2.3.0
